Rename confusing double-use of the name util.h in rcx
There is a util.h in the includes of rcx, but also a util.{h,cpp}
in the src part; different content, different utils.
Rename the one inside src/ to what it contains (findSomeNet() fun).

While at it, start using project-relative paths for these.
