+++ src/rcx/BUILD
@@ -85,8 +85,8 @@ cc_library(
         "src/parse.cpp",
         "src/parse.h",
         "src/process_ext.cpp",
-        "src/util.cpp",
-        "src/util.h",
+        "src/find_some_net.cpp",
+        "src/find_some_net.h",
     ],
     includes = [
         "include",
