+++ src/rcx/src/find_some_net.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Copyright (c) 2019-2025, The OpenROAD Authors
 
-#include "util.h"
+#include "find_some_net.h"
 
 #include <memory>
 #include <vector>
